feat: 팀이 없을 경우 페이지 구현 및 기존 페이지 이동
